+++ htdocs/inc/content/drugs/drug/inc.drugs.drug.acetylsalicylsaeure.php
@@ -25,6 +25,7 @@
 				<h2>Wirkung</h2>
 				<p>Acetylsalicylsäure wirkt durch Hemmung der Cyclooxygenase COX-1 schon in kleinen Dosen (30–50 mg) gerinnungshemmend, mit steigender Dosis (0,5–2 g) durch Hemmung der Cyclooxygenase COX-1 und COX-2 und der daraus folgenden sinkenden Bildung von Prostaglandinen auch dezentral schmerzstillend, antirheumatisch sowie fiebersenkend und schließlich (2–5 g) entzündungshemmend.</p>
 			</div>
+			<!-- !Nebenwirkung-->
 			<div data-role="collapsible" name="mediSortItem3">
 				<h2>Nebenwirkung</h2>
 				<p>Höhere Dosen (ab 6-8 g/Tag) von ASS haben eine Ototoxizität und können das Gehör schädigen und Tinnitus verursachen.</p>
@@ -66,17 +67,82 @@
 				</ul>
 				<p>Acetylsalicylsäure zerfällt bei zu langer Lagerung zu Salicylsäure und führt dann zu verstärkten Nebenwirkungen.</p>
 			</div>
+			<!-- !Interaktion/Wechselwirkungen -->
 			<div data-role="collapsible" name="mediSortItem4">
 				<h2>Interaktion &frasl; Wechselwirkungen</h2>
+				<ul>
+					<li>
+						<dl>
+							<dt>Antikoagulanzien (blutgerinnungshemmende Mittel)</dt>
+							<dd>Erhöhung der Blutungsgefahr (gastrointerstinal). Besonders gefährlich sind Kombinationen mit Clopidogrel und Vitamin-K-Antagonisten (Warfarin, Dicoumarol).</dd>
+						</dl>
+					</li>
+					<li>
+						<dl>
+							<dt>Antidiabetika</dt>
+							<dd>Verstärkt die Wirkung der Antidiabetika</dd>
+						</dl>
+					</li>
+					<li>
+						<dl>
+							<dt>Schleifendiuretika, Spironolacton, Canrenoat, Blutdrucksenker und Gichtmittel</dt>
+							<dd>Vermindert die Wirkung dieser Medikamente</dd>
+						</dl>
+					</li>
+					<li>
+						<dl>
+							<dt>Digoxin, Lithium oder Barbiturate</dt>
+							<dd>Konzentration im Blut wird erhöht. Erhöht Vergiftungsrisiko.</dd>
+						</dl>
+					</li>
+					<li>
+						<dl>
+							<dt>Methotrexat (Zellgift zur Behandlung von Rheuma und Krebs)</dt>
+							<dd>Das Schilddrüsenhormon Trijodthyronin oder Sulfonamid werden in Wirkung und Nebenwirkungen verstärkt</dd>
+						</dl>
+					</li>
+					<li>
+						<dl>
+							<dt>Valproinsäure (Antiepileptikum)</dt>
+							<dd>Kann zur Verstärkung der Wirkung und Nebenwirkung führen.</dd>
+						</dl>
+					</li>
+					<li>
+						<dl>
+							<dt>alle nicht-steroidalen Antirheumatika</dt>
+							<dd>Konzentration im Blut wird erhöht, dies kann zur Steigerung der Wirkung und Nebenwirkung führen.</dd>
+						</dl>
+					</li>
+				</ul>	
 			</div>
 			<div data-role="collapsible" name="mediSortItem5">
 				<h2>Wirkungseintritt</h2>
+				<dl>
+					<dt>Thrombozytenaggregationshemmer</dt>
+					<dd>Wirkungseintritt bereits nach 2 Minuten</dd>
+				</dl>
+				<dl>
+					<dt>Schmerzen</dt>
+					<dd>Wirkungseintritt nach 4-10 Minuten - Wirkungsmaximum nach 20 Minuten</dd>
+				</dl>
 			</div>
 			<div data-role="collapsible" name="mediSortItem6">
 				<h2>Wirkungsdauer</h2>
+				<dl>
+					<dt>Thrombozytenaggregationshemmer</dt>
+					<dd>Wirkungsdauer 3-5 Tage nach Einnahme&frasl;Gabe</dd>
+				</dl>
+				<dl>
+					<dt>Schmerzen</dt>
+					<dd>Wirkungsdauer 3-4 Stunden nach Einnahme&frasl;Gabe</dd>
+				</dl>
 			</div>
 			<div data-role="collapsible" name="mediSortItem7">
-				<h2>Zusammensetzung</h2>
+				<h2>Darreichungsformen</h2>
+				<ul>
+					<li>Kautablette</li>
+					<li>Injektionsflaschen (Pulver)</li>
+				</ul>
 			</div>
 			<div data-role="collapsible" name="mediSortItem8">
 				<h2>Indikation</h2>
@@ -86,34 +152,56 @@
 			</div>
 			<div data-role="collapsible" name="mediSortItem10">
 				<h2>Dosierungsempfehlung</h2>
+				<table data-role="table" id="table-custom-2" data-mode="columntoggle" class="ui-body-d ui-shadow table-stripe ui-responsive" data-column-btn-theme="b" data-column-btn-text="Kategorie" data-column-popup-theme="a">
+					<thead>
+						<tr>
+							<th>Situation</th>
+							<th data-priority="1">Erwachsen</th>
+							<th data-priority="2">Kinder</th>
+						</tr>
+					</thead>
+					<tbody>
+						<tr>
+							<th>Akutes Koronarsyndrom</th>
+							<td>Die Initialdosis von ASS-Kautabletten beträgt 160-325 mg. Andere Formen von ASS (löslich, i.v.-Präparation) sind wohl ebenso effektiv.</td>
+							<td>nicht bekannt</td>
+						</tr>
+						<tr>
+							<th>Akuter Migräneanfall</th>
+							<td></td>
+							<td></td>
+						</tr>
+						<tr>
+							<th>Schmerzen und Fieber</th>
+							<td></td>
+							<td></td>
+						</tr>
+					</tbody>
+				</table>
 			</div>
 			<div data-role="collapsible" name="mediSortItem11">
 				<h2>bek. Handelsnamen</h2>
+				<p>Aspisol, Aspirin, ASS, </p>
 			</div>
 		</div>
 		<h2>Quellenangaben</h2>
 		<ul>
 			<li>Online
 				<ol>
+					<li><a href="#" onclick="openDeviceBrowser('http://www.grc-org.de/vortraege/cat_view/6-wissenschaft/1-leitlinien/27-leitlinien-2010');">Deutscher Rat für Wiederbelebung - German Resuscitation Council (GRC) e.V. (ERC Leitlinien 2010 in Deutsch)</a></li>
 					<li><a href="#" onclick="openDeviceBrowser('http://flexikon.doccheck.com/Acetylsalicylsäure');">DocCheck Flexikon</a>, das Medizinlexikon zum Mitmachen (<a href="#" onclick="openDeviceBrowser('http://flexikon.doccheck.com/de/Spezial:Artikel_Autoren/Acetylsalicylsäure');">Liste der Autoren</a>)</li>
 					<li><a href="#" onclick="openDeviceBrowser('http://de.wikipedia.org/wiki/Acetylsalicylsäure');">Wikipedia</a></li>
+					<li><a href="#" onclick="openDeviceBrowser('http://www.onmeda.de/Wirkstoffe/Acetylsalicyls%E4ure/');">Onmeda</a></li>
 				</ol>
 			</li>
 			<li>Literatur
 				<ol>
-					<li></li>
+					<li>Böhmer, Roman &frasl; Schneider, Dr. med. Thomas H. &frasl; Wolcke, Dr. med. Benno: Taschenatlas Rettungsdienst, Gau-Bischofsheim: Naseweis Verlag, 10., aktualisierte Auflage. Auflage 2013, Seite 572-573</li>
+					<li>Ruß, Dr. med. Andreas: Arzneimittel pocket, Grünwald: Börm Bruckmeier Verlag GmbH, Sonderauflageauflage Oktober 2012, Seite 14, 29, 95-96, 120, 128, 367</li>
 				</ol>
 			</li>
 		</ul>
 	</div>
-	<!--<div data-role="footer" data-position="fixed">
-		<div data-role="navbar">
-			<ul>
-				<li><a href="#drugs-drug-acetylsalicylsaeure" data-icon="info">Generell</a></li>
-				<li><a href="#drugs-drug-acetylsalicylsaeure-regional" data-icon="location">Regional</a>
-			</ul>
-		</div>
-	</div>-->
 </div>
 <div data-role="page" id="drugs-drug-acetylsalicylsaeure-regional" data-quicklinks="true">
 	<div data-role="header" data-position="fixed">