+++ htdocs/inc/content/main/inc.main.03.startup.php
@@ -1,0 +1,17 @@
+<div data-role="page" id="main-03-startup">
+	<div data-role="content" role="main" class="ui-content">
+		<div class="container">
+			<div id="slides">
+				<img src="core/img/startupslide/buttonsettings_700x700.jpg" />
+				<img src="core/img/startupslide/settingsupdate_700x700.jpg" />
+				<img src="core/img/startupslide/buttonhome_700x700.jpg" />
+				<img src="core/img/startupslide/buttonmenu_700x700.jpg" />
+				<a href="#" class="slidesjs-previous slidesjs-navigation"><i class="icon-chevron-left icon-large"></i></a>
+				<a href="#" class="slidesjs-next slidesjs-navigation"><i class="icon-chevron-right icon-large"></i></a>
+			</div>
+		</div>
+		<p>
+			<a href="#main-02-index" class="ui-btn ui-icon-carat-r ui-btn-icon-right ui-corner-all">Slider überspringen...</a>
+		</p>
+	</div>
+</div>