+++ htdocs/inc/content/other/inc.other.01.donate.php
@@ -1,4 +1,4 @@
-<div data-role="page" id="other-01-donate" data-quicklinks="true">
+<div data-role="page" id="other-01-donate">
 	<div data-role="header" data-position="fixed">
 		<div data-role="controlgroup" data-type="horizontal" class="ui-mini ui-btn-left">
 			<a href="#function-99-navigationpanel" class="ui-btn ui-btn-icon-right ui-icon-bars ui-btn-icon-notext">Menü</a>
@@ -13,8 +13,8 @@
 	<div data-role="content" role="main" class="ui-content">
 		<p>Diese App ist von Rettungsdienstler für Rettungsdienstler - uns ist wichtig, dass alle Informationen kostenlos und frei zugänglich für Sie zur Verfügung gestellt werden! Das Entwickeln/Bereitstellen ist leider nicht ohne finanziellen Aufwand möglich. Sollte Ihnen die RettApp gefallen so unterstützen Sie den Entwickler!</p>
 		<p class="align-center">
-			<a href="http://www.amazon.de/registry/wishlist/YRKA7VXAPXGA" data-rel="external" target="foo"><img src="core/icons/1391581671_Amazon.png" alt="1391581671_Amazon" width="120" /></a>
-			<a href="https://www.paypal.com/cgi-bin/webscr?cmd=_s-xclick&amp;hosted_button_id=G62QH4JXWWUKU" data-rel="external" target="foo"><img src="core/icons/1391581724_PayPal.png" alt="1391581724_PayPal" width="120" /></a>
+			<a href="http://www.amazon.de/registry/wishlist/YRKA7VXAPXGA" data-rel="external" target="foo"><img src="core/icons/1392833874_Amazon.png" alt="Amazon Button" width="120" /></a>
+			<a href="https://www.paypal.com/cgi-bin/webscr?cmd=_s-xclick&amp;hosted_button_id=G62QH4JXWWUKU" data-rel="external" target="foo"><img src="core/icons/1392833865_PayPal.png" alt="PayPal Button" width="120" /></a>
 		</p>
 	</div>
 </div>