+++ htdocs/inc/content/other/inc.other.03.releasenotes.php
@@ -1,0 +1,142 @@
+<div data-role="page" id="other-03-releasenotes">
+	<div data-role="header" data-position="fixed">
+		<div data-role="controlgroup" data-type="horizontal" class="ui-mini ui-btn-left">
+			<a href="#function-99-navigationpanel" class="ui-btn ui-btn-icon-right ui-icon-bars ui-btn-icon-notext">Menü</a>
+			<a data-rel="back" class="ui-btn ui-btn-icon-right ui-icon-arrow-l ui-btn-icon-notext">Zurück</a>
+		</div>
+		<h1>Release Notes</h1>
+		<div data-role="controlgroup" data-type="horizontal" class="ui-mini ui-btn-right">
+			<a href="#function-02-settings" class="ui-btn ui-btn-icon-right ui-icon-gear ui-btn-icon-notext" data-transition="slidefade" data-direction="reverse">Einstellungen</a>
+			<a href="#main-02-index" class="ui-btn ui-btn-icon-right ui-icon-home ui-btn-icon-notext">Home</a>
+		</div>
+	</div>
+	<div data-role="content" role="main" class="ui-content">
+		<div data-role="collapsibleset" data-theme="a" data-content-theme="a">
+			<div data-role="collapsible">
+				<h2>Version 0.0.3<span class="ui-li-count">Aktuell</span></h2>
+				<ul>
+					<li><strong>Wichtig!</strong> Änderungen aller Nummern von Vorwahl-112 zu Vorwahl-19222 (Danke Annika, Patric, Thomas)</li>
+					<li>jQuery Mobile Update auf 1.4.2</li>
+					<li>Update auf Phonegap 3.3.0 - dies könnte Probleme lösen</li>
+					<li>Lade Fehler bei schlechtem Internet behoben</li>
+					<li>Splash-Screens wurden hinzugefügt</li>
+					<li>Dosierer:
+						<ul>
+							<li>Neuer Name, nicht mehr Dosierrechner</li>
+							<li>Neues Format</li>
+						</ul>
+					</li>
+					<li>Standort:
+						<ul>
+							<li>Verfolgung heißt jetzt Karte zentrieren (Genauere Definition der Funktion)</li>
+							<li>Leitstellen nun mit Funkrufnamen und Funkart (Digital/Analog)</li>
+							<li>Neue Leitstellen
+								<ul>
+									<li>Florian Lübeck</li>
+									<li>Leitstelle Neumünster</li>
+									<li>Leitstelle West: Dithmarschen</li>
+									<li>Leitstelle Süd: Herzogtum Lauenburg, Ostholstein</li>
+									<li>Leitstelle Nord: Nordfriesland</li>
+									<li>Leitstelle Bergstraße</li>
+									<li>Leitstelle Wesel</li>
+									<li>Florian Viersen</li>
+									<li>Leitstelle Bonn</li>
+									<li>Leitstelle Rhein-Sieg</li>
+									<li>Florian Leverkusen</li>
+									<li>Florian Erft</li>
+									<li>Leitstelle Euskirchen</li>
+									<li>Florian Bottrop</li>
+									<li>Florian Heinsberg</li>
+									<li>Leitstelle Oberberg</li>
+									<li>Leitstelle Berga</li>
+									<li>Florian Gelsenkirchen</li>
+									<li>Florian Münster</li>
+									<li>Florian Borken</li>
+									<li>Florian Coesfeld</li>
+									<li>Florian Recklinghausen</li>
+									<li>Florian Steinfurt</li>
+									<li>Florian Warendorf</li>
+									<li>Leitstelle Bielefeld</li>
+									<li>Florian Gütersloh</li>
+									<li>Leitstelle Herford</li>
+									<li>Leitstelle Kreis Höxter</li>
+									<li>Leitstelle Minden</li>
+									<li>Florian Bochum</li>
+									<li>Leitstelle Dortmund</li>
+									<li>Florian Hagen</li>
+									<li>Florian Hamm</li>
+									<li>Florian Herne</li>
+									<li>Leitstelle Ennepe</li>
+									<li>Leitstelle Sauerland</li>
+								</ul>
+							</li>
+						</ul>
+					</li>
+					<li>Hauptfunktion Medikamente wird eingeführt
+						<ul>
+							<li>erstes Medikament ASS als Beispiel eingefügt (noch nicht Komplett)</li>
+						</ul>
+					</li>
+					<li>Einstellung:
+						<ul>
+							<li>Sortierung der Informationen von Medikamenten</li>
+							<li>Disclaimer umbenannt in Rechtliches</li>
+							<li>Neue Leitstellen:
+								<ul>
+									<li>Baden-Württemberg (BW)</li>
+									<li>Bayern (BY)</li>
+									<li>Berlin (BE)</li>
+									<li>Brandenburg (BB)</li>
+									<li>Bremen (BRE)</li>
+									<li>Hamburg (HH)</li>
+									<li>Hessen (HE)</li>
+									<li>Saarland (SL)</li>
+								</ul>
+							</li>
+						</ul>
+					</li>
+					<li>Telefonnummern Leitstelle Mainz korriegiert (Danke Björn)</li>
+					<li>Telefonnummern von Ludwigshafen korrigiert (Danke lafidi @ community.rettapp.org)</li>
+					<li>Sponsoren:
+						<ul>
+							<li>Amazon-Button korrigiert</li>
+							<li>PayPal-Button korrigiert</li>
+						</ul>
+					</li>
+				</ul>
+			</div>
+			<div data-role="collapsible">
+				<h2>Version 0.0.2</h2>
+				<ul>
+					<li>Fehler Medikamenten Disclaimer behoben (Danke Jörn)</li>
+					<li>Versionscheck eingepflegt</li>
+					<li>Telefonnummer Leitstelle Mainz korrigiert</li>
+					<li>Rechtschreibfehler korrigiert (Danke Lisa)</li>
+					<li>Neue Leitstellen in Standort:
+						<ul>
+							<li>Flensburg</li>
+							<li>Kiel</li>
+							<li>Rheingau-Taunus-Kreis</li>
+							<li>Wiesbaden</li>
+							<li>Main-Taunus-Kreis</li>
+							<li>Groß-Gerau</li>
+							<li>Hochtaunuskreis</li>
+							<li>ILS Rhein-Neckar</li>
+							<li>Karlsruhe</li>
+						</ul>
+					</li>
+					<li>Copyright neu ausgebaut</li>
+					<li>Icons optimiert</li>
+					<li>Medikamente -> Krankheiten vorläufig entfernt</li>
+					<li>Medikamente -> Regional (Freigaben) hinzugefügt</li>
+				</ul>
+			</div>
+			<div data-role="collapsible">
+				<h2>Version 0.0.1</h2>
+				<ul>
+					<li>initial Release</li>
+				</ul>
+			</div>
+		</div>
+	</div>
+</div>