+++ htdocs/Release_notes.md
@@ -24,27 +24,73 @@ Version 002:
 Version 003:
 ============
 
-- Wichtig: Änderungen aller Nummern von Vorwahl-112 zu Vorwahl-19222 (Danke Annika + Patric)
-- Dosierrechner jetzt Dosierer
-- Dosierer öffnet im neuen Format
-- Leitstellen in Standort nun mit Funkrufnamen und Funkart (Digital/Analog)
-- Neue Leitstellen in Standort:
--- Florian Lübeck
--- Leitstelle Neumünster
--- Leitstelle West: Dithmarschen
--- Leitstelle Süd: Herzogtum Lauenburg, Ostholstein
--- Leitstelle Nord: Nordfriesland
-
-- Neue Leitstellen in den Einstellungen:
--- Baden-Württemberg (BW)
--- Bayern (BY)
--- Berlin (BE)
--- Brandenburg (BB)
--- Bremen (BRE)
--- Hamburg (HH)
--- Hessen (HE)
--- Saarland (SL)
-
+- Wichtig!</strong> Änderungen aller Nummern von Vorwahl-112 zu Vorwahl-19222 (Danke Annika, Patric, Thomas)
+- jQuery Mobile Update auf 1.4.2
+- Update auf Phonegap 3.3.0 - dies könnte Probleme lösen
+- Lade Fehler bei schlechtem Internet behoben
+- Splash-Screens wurden hinzugefügt
+- Dosierer:
+- - Neuer Name, nicht mehr Dosierrechner
+- - Neues Format
+- Standort:
+- - Verfolgung heißt jetzt Karte zentrieren (Genauere Definition der Funktion)
+- - Leitstellen nun mit Funkrufnamen und Funkart (Digital/Analog)
+- - Neue Leitstellen
+- - - Florian Lübeck
+- - - Leitstelle Neumünster
+- - - Leitstelle West: Dithmarschen
+- - - Leitstelle Süd: Herzogtum Lauenburg, Ostholstein
+- - - Leitstelle Nord: Nordfriesland
+- - - Leitstelle Bergstraße
+- - - Leitstelle Wesel
+- - - Florian Viersen
+- - - Leitstelle Bonn+
+- - - Leitstelle Rhein-Sieg
+- - - Florian Leverkusen
+- - - Florian Erft
+- - - Leitstelle Euskirchen
+- - - Florian Bottrop
+- - - Florian Heinsberg
+- - - Leitstelle Oberberg
+- - - Leitstelle Berga
+- - - Florian Gelsenkirchen
+- - - Florian Münster
+- - - Florian Borken
+- - - Florian Coesfeld
+- - - Florian Recklinghausen
+- - - Florian Steinfurt
+- - - Florian Warendorf
+- - - Leitstelle Bielefeld
+- - - Florian Gütersloh
+- - - Leitstelle Herford
+- - - Leitstelle Kreis Höxter
+- - - Leitstelle Minden
+- - - Florian Bochum
+- - - Leitstelle Dortmund
+- - - Florian Hagen
+- - - Florian Hamm
+- - - Florian Herne
+- - - Leitstelle Ennepe
+- - - Leitstelle Sauerland
 - Hauptfunktion Medikamente wird eingeführt
-- Einstellung: Sortierung der Informationen von Medikamenten
-- Telefonnummern von Ludwigshafen korregiert (Danke lafidi @ community@rettapp.org)
+- - erstes Medikament ASS als Beispiel eingefügt (noch nicht Komplett)
+- Einstellung:
+- - Sortierung der Informationen von Medikamenten
+- - Disclaimer umbenannt in Rechtliches
+- - Neue Leitstellen:
+- - - Baden-Württemberg (BW)
+- - - Bayern (BY)
+- - - Berlin (BE)
+- - - Brandenburg (BB)
+- - - Bremen (BRE)
+- - - Hamburg (HH)
+- - - Hessen (HE)
+- - - Saarland (SL)
+- Telefonnummern Leitstelle Mainz korriegiert (Danke Björn)
+- Telefonnummern von Ludwigshafen korrigiert (Danke lafidi @ community.rettapp.org)
+- Sponsoren:
+- - Amazon-Button korrigiert
+- - PayPal-Button korrigiert
+						
+					
+